Migrate `st.image` e2e test to playwright
